requirements.txt
altair==4.2.0
cryptography==37.0.1
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.3
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.8.1
seaborn==0.11.2
